requirements.txt
Django==1.11.8
django-oauth-toolkit==1.0.0
djangorestframework==3.7.7